backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
paho-mqtt==2.1.0
python-dotenv==1.0.1
tzdata==2025.1
